Updated proj to compile with CCC 5.2 on linux...almost works
Had to switch all the cutil*.h includes to helper_*.h, and change
a bunch of method names that were changed since the days of CCC 2.0.
Also removed the (huge) common directory, and replaced with a hard
path to /usr/local/cuda/samples/common, which appears to be where
Cuda SDK 7.0+ is installed.

With my GTX 960, it does compile and run, but there's no image,
indicating that the CUDA-OpenGL interop system has probably changed
since 2011.  This should still be usable for generating still pngs.